add the ability to inject data
This is just wrong, but it's the first pass of where I actually should belong
which is to expose this internall only. Verse should provide the API that
endpoints can use to fetch this data themselves before editing the page.

The real fix for this will include making page.data const so once a page is
'prepared' it can be modified.

My only regret is that I didn't see that before now... fix coming