mov more of the template engine into comptime
This is a partial commit, while I decide if I wanted to write the larger
testing template (this compiles and passes tests, but srctree won't
compile because a lot of the "recursive" template code isn't finished
yet.

I might also end up rewriting the directive code instead, see the
commented out `page` line, where it might make more sense for this to be
a generated type. Another option is to turn directive into a union
itself, and isolate each type into it's own struct. Still unsure what
the cleanest is likely to be.